MB-59470: In case of Delete Only TombStone is sent to Kafka
Motivation:

When a record is deleted from dynamo side, kafka topic
received two records, first image is of previous one while
the second image contains the null value. The first image is
not needed as based on the second image entry will be deleted
as the value of entry is null.

The issue it was causing was in case of batching on sink connector
side, since first and second image contains same key the batcher
put them in seperate batch, so if multiple delete starts coming
up where every entry will have two images on topic side, then the
batch will create 2*x delete requests instead of batch all delete
in just 1 request.

Change-Id: I713c9b648fa44b7af37efcb2c09ca7125fa0a6ba
Reviewed-on: https://review.couchbase.org/c/kafka-connect-dynamodb/+/200174
Tested-by: Ritik Raj <raj.ritik9835@gmail.com>
Reviewed-by: <utsav.singh@couchbase.com>